# Simplify __invert__ for polynomial quotient ring element