@vincentmacri 2025-01-29
@user202729 I'm curious if this change has any performance impact. Would you mind posting a before/after `timeit` result?