@user202729 2025-01-30
That's a good question… on the other hand we can argue `inverse_mod` should be exactly equivalent to quotient ring element `__invert__`, otherwise one can be optimized in term of the other.

On the other hand, what happened previously is that they both have different supported/unsupported cases (e.g. previously `inverse_mod` doesn't fallback to `inverse_of_unit`, while `__invert__` does). When they're merged, it seems best to try to support as many cases as possible, this necessarily slows down the operator because more cases need to be checked.

There's the caveat of what to do with inexact rings… we have to adopt one option or the other either way.