@vincentmacri 2025-01-30
From my reading of this you added some extra checking in `polynomial_element.pyx`, and changed `__invert__` in `polynomial_quotient_ring_element.py` to call the function in `polynomial_element.pyx`? I expect that from this change the performance hit in `polynomial_element.pyx` would be negligible, and `__invert__` in `polynomial_quotient_ring_element.py` will be sped up since it's not calling a function written in Cython.

I would like to see before/after `timeit` results of the two functions before approving this, just in case the reason for the two different implementations was because it had some performance benefits.

Additionally, your PR description mentions that you fixed a few bugs with this PR. Please add tests covering those bugs if you haven't yet, and add a pointer to this PR in the explanation of those tests (i.e. "See :issue:`39402`").