@user202729 2025-01-31
Not exactly bug, just missing feature — that previously `inverse_mod` didn't fallback to `inverse_of_unit`, but it now does (see the newly added doctest).

Also, I don't think there's runtime overhead of passing between the Python–Cython boundary, except for what is already there in Python interpreter.

Timing result:

Setup code:

```sage
sage: R.<x> = GF(5)[]
....: S = R.quotient(x^3+1)
sage: s_element = S(x+2)
....: r_element = x+2
....: modulus = x^3+1
```

Before the PR:

```sage
sage: %timeit ~s_element
11.9 μs ± 341 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage: %timeit r_element.inverse_mod(modulus)
4.73 μs ± 185 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

After the PR:

```sage
sage: %timeit ~s_element
9.72 μs ± 261 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
sage: %timeit r_element.inverse_mod(modulus)
4.81 μs ± 267 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

So… `__invert__` got sped up, and `inverse_mod` gets slightly slowed down (understandable because the additional check to fallback to `inverse_of_unit` is needed)