@vincentmacri 2025-01-31
Great! Looks good to me. The `inverse_mod` slowdown looks like it's probably within the range of measurement error anyway.